refactor(web): rename live-voice preflight util and consolidate verdict type
